+++ redux-1/src/types/agenda.js
